+++ src/main/java/depth/main/ideac/domain/mail/application/MailService.java
@@ -3,7 +3,7 @@ package depth.main.ideac.domain.mail.application;
 import depth.main.ideac.domain.mail.domain.Verify;
 import depth.main.ideac.domain.mail.domain.repository.MailRepository;
 import depth.main.ideac.domain.mail.dto.FindPasswordReq;
-import depth.main.ideac.domain.mail.dto.PasswordReq;
+import depth.main.ideac.domain.user.dto.PasswordReq;
 import depth.main.ideac.domain.user.domain.User;
 import depth.main.ideac.domain.user.domain.repository.UserRepository;
 import depth.main.ideac.global.DefaultAssert;
@@ -79,34 +79,5 @@ public class MailService {
         return mailRepository.save(new Verify(code,email, LocalDateTime.now().plusMinutes(3)));
     }
 
-    @Transactional
-    public ResponseEntity<?> changePassword(@Valid PasswordReq passwordReq, String code){
 
-        //검증
-        DefaultAssert.isTrue(passwordReq.getPassword().equals(passwordReq.getRePassword()), "비밀번호가 서로 다릅니다.");
-        //만료시간 검증
-        Verify verify = mailRepository.findByCode(code);
-
-        if (verify == null){
-            throw new DefaultException(ErrorCode.INVALID_CHECK, "이미변경되었습니다?");
-        }
-
-        DefaultAssert.isTrue(verify.checkExpiration(LocalDateTime.now()), "만료되었습니다.");
-
-        Optional<User> findUser = userRepository.findByEmail(verify.getEmail());
-
-        User user = findUser.get();
-        user.updatePassWord(passwordEncoder.encode(passwordReq.getPassword()));
-
-        // 인증완료 후 삭제
-        mailRepository.delete(verify);
-
-        ApiResponse apiResponse = ApiResponse.builder()
-                .check(true)
-                .information(null)
-                .message("비밀번호를 바꾸었어요")
-                .build();
-
-        return ResponseEntity.ok(apiResponse);
-    }
 }
